pytest_plugin: Speedup teardown of cassandra fixture significantly
Truncate a keyspace table only if it is not empty when executing the
teardown phase of the swh_storage_cassandra_backend_config function
scope fixture.

This brings a two times speedup when executing all cassandra related
tests.

+++ swh/storage/pytest_plugin.py
@@ -263,7 +263,11 @@ def swh_storage_cassandra_backend_config(
     storage = get_storage(**storage_config)
 
     for table in TABLES:
-        storage._cql_runner._session.execute(f"TRUNCATE TABLE {keyspace}.{table}")
+        table_rows = storage._cql_runner._session.execute(
+            f"SELECT * from {keyspace}.{table} LIMIT 1"
+        )
+        if table_rows.one() is not None:
+            storage._cql_runner._session.execute(f"TRUNCATE TABLE {keyspace}.{table}")
 
     storage._cql_runner._cluster.shutdown()
